Fix view component
I saw [your blog post](https://www.compositional-it.com/news-blog/stateful-react-components/), and noticed this minor issue. When using the `React.functionComponent` syntax , components need to not be a function. This is because the internal state (if the menu is open) will be reset when the function is evaluated (when the model changes).